Remove use of `RefCell`
This change asks the closure in the argument of `Subgrid::convolute` to
be `&mut dyn FnMut` instead of `&dyn Fn`. The latter does not allow
modifications to its captured environment, which is why a `RefCell` was
needed to work around this limitation. Using an `FnMut` makes this much
simpler as it doesn't have this limitation.