Improve evolution performance for double-hadronic grids
Instead of using a cartesian product of two PIDs which is subsequently
filtered use two iterators of PIDs first filtered and subsequently
zipped. This is a change of O(n^2) to O(n) which makes a big difference
if n is large enough. This is the case for dijet grids, for example.